--- test/e2e/tests/settings/about-metamask-ui-validation.spec.ts
@@ -1,47 +1,0 @@
-import { Suite } from 'mocha';
-import FixtureBuilderV2 from '../../fixtures/fixture-builder-v2';
-import { withFixtures } from '../../helpers';
-import { Driver } from '../../webdriver/driver';
-import packageJson from '../../../../package.json';
-import AboutPage from '../../page-objects/pages/settings/about-page';
-import HeaderNavbar from '../../page-objects/pages/header-navbar';
-import HomePage from '../../page-objects/pages/home/homepage';
-import SettingsPage from '../../page-objects/pages/settings/settings-page';
-import { login } from '../../page-objects/flows/login.flow';
-import { closeSettings } from '../../page-objects/flows/settings.flow';
-
-// Test case to validate the view in the "About" - MetaMask.
-describe('Setting - About MetaMask :', function (this: Suite) {
-  it('validate the view', async function () {
-    await withFixtures(
-      {
-        fixtures: new FixtureBuilderV2().build(),
-        title: this.test?.fullTitle(),
-      },
-      async ({ driver }: { driver: Driver }) => {
-        await login(driver);
-
-        // navigate to settings and click on about page
-        await new HeaderNavbar(driver).openSettingsPage();
-        const settingsPage = new SettingsPage(driver);
-        await settingsPage.checkPageIsLoaded();
-        await settingsPage.goToAboutPage();
-
-        const aboutPage = new AboutPage(driver);
-        await aboutPage.checkPageIsLoaded();
-
-        // verify the version number of MetaMask
-        const { version } = packageJson;
-        await aboutPage.checkMetaMaskVersionNumber(version);
-
-        // click on `close` button
-        await closeSettings(driver);
-
-        // wait for home page and validate the balance
-        const homePage = new HomePage(driver);
-        await homePage.checkPageIsLoaded();
-        await homePage.checkExpectedBalanceIsDisplayed();
-      },
-    );
-  });
-});
