--- test/e2e/tests/settings/address-book.spec.ts
@@ -1,290 +1,0 @@
-import { Suite } from 'mocha';
-import { Mockttp } from 'mockttp';
-import { withFixtures } from '../../helpers';
-import { shortenAddress } from '../../../../ui/helpers/utils/util';
-import FixtureBuilderV2 from '../../fixtures/fixture-builder-v2';
-import HomePage from '../../page-objects/pages/home/homepage';
-import ActivityTab from '../../page-objects/pages/home/activity-tab';
-import ContactsPage from '../../page-objects/pages/settings/contacts-settings';
-import HeaderNavbar from '../../page-objects/pages/header-navbar';
-import TransactionConfirmation from '../../page-objects/pages/confirmations/transaction-confirmation';
-import { login } from '../../page-objects/flows/login.flow';
-import NetworkManager from '../../page-objects/pages/network-manager';
-import { TOKENS_API_MOCK_RESULT } from '../../../data/mock-data';
-import { createInternalTransaction } from '../../page-objects/flows/transaction.flow';
-import { NETWORK_CLIENT_ID } from '../../constants';
-
-async function mockTokenList(mockServer: Mockttp) {
-  return await mockServer
-    .forGet('https://bridge.api.cx.metamask.io/networks/59144/tokens')
-    .thenCallback(() => {
-      return {
-        statusCode: 200,
-        json: TOKENS_API_MOCK_RESULT,
-      };
-    });
-}
-
-describe('Address Book', function (this: Suite) {
-  it('Sends to an address book entry', async function () {
-    await withFixtures(
-      {
-        fixtures: new FixtureBuilderV2()
-          .withAddressBookController({
-            addressBook: {
-              '0x539': {
-                '0x2f318C334780961FB129D2a6c30D0763d9a5C970': {
-                  address: '0x2f318C334780961FB129D2a6c30D0763d9a5C970',
-                  chainId: '0x539',
-                  isEns: false,
-                  memo: '',
-                  name: 'Test Name 1',
-                },
-              },
-            },
-          })
-          .build(),
-        title: this.test?.fullTitle(),
-      },
-      async ({ driver }) => {
-        await login(driver);
-
-        // Add flakiness fix here: wait for the continue button to be stably enabled
-        await createInternalTransaction({
-          driver,
-          chainId: '0x539',
-          symbol: 'ETH',
-          recipientName: 'Test Name 1',
-          amount: '2',
-        });
-
-        await new TransactionConfirmation(driver).clickFooterConfirmButton();
-
-        const homePage = new HomePage(driver);
-        await homePage.goToActivityList();
-        const activityTab = new ActivityTab(driver);
-        await activityTab.checkConfirmedTxNumberDisplayedInActivity(1);
-        await activityTab.checkTxAction({ action: 'Sent ETH' });
-        await activityTab.checkTxAmountInActivity(`-2 ETH`, 1);
-      },
-    );
-  });
-
-  it('Sends to an address book entry on a different network', async function () {
-    await withFixtures(
-      {
-        fixtures: new FixtureBuilderV2()
-          .withSelectedNetwork(NETWORK_CLIENT_ID.LINEA_MAINNET)
-          .withAddressBookController({
-            addressBook: {
-              '0x539': {
-                '0x2f318C334780961FB129D2a6c30D0763d9a5C970': {
-                  address: '0x2f318C334780961FB129D2a6c30D0763d9a5C970',
-                  chainId: '0x539',
-                  isEns: false,
-                  memo: '',
-                  name: 'Test Name 1',
-                },
-              },
-            },
-          })
-          .withEnabledNetworks({
-            eip155: {
-              '0xe708': true,
-            },
-          })
-          .build(),
-        testSpecificMock: mockTokenList,
-        title: this.test?.fullTitle(),
-      },
-      async ({ driver }) => {
-        // Start on Linea
-        await login(driver);
-
-        // Send transaction on Localhost
-        await createInternalTransaction({
-          driver,
-          chainId: '0x539',
-          symbol: 'ETH',
-          recipientName: 'Test Name 1',
-          amount: '2',
-        });
-
-        const confirmation = new TransactionConfirmation(driver);
-        await confirmation.waitForReviewAlertToDisappear();
-        await confirmation.clickFooterConfirmButton();
-
-        // Select Linea to check the Activity list
-        const networkSelector = new NetworkManager(driver);
-        await networkSelector.openNetworkManager();
-        await networkSelector.selectNetworkByName('Localhost 8545');
-
-        const homePage = new HomePage(driver);
-        await homePage.goToActivityList();
-        const activityTab = new ActivityTab(driver);
-        await activityTab.checkConfirmedTxNumberDisplayedInActivity(1);
-        await activityTab.checkTxAction({ action: 'Sent ETH' });
-        await activityTab.checkTxAmountInActivity(`-2 ETH`, 1);
-      },
-    );
-  });
-
-  it('Adds a new contact to the address book', async function () {
-    await withFixtures(
-      {
-        fixtures: new FixtureBuilderV2().build(),
-        title: this.test?.fullTitle(),
-      },
-      async ({ driver }) => {
-        await login(driver);
-
-        await new HeaderNavbar(driver).openContactsPage();
-        const contactsPage = new ContactsPage(driver);
-        await contactsPage.checkPageIsLoaded();
-        await contactsPage.addContact(
-          'Test User',
-          '0x56A355d3427bC2B1E22c78197AF091230919Cc2A',
-        );
-        await contactsPage.checkContactDisplayed({
-          contactName: 'Test User',
-          address: shortenAddress('0x56A355d3427bC2B1E22c78197AF091230919Cc2A'),
-        });
-      },
-    );
-  });
-
-  it('Adds a new contact to the address book on a different chain', async function () {
-    await withFixtures(
-      {
-        fixtures: new FixtureBuilderV2().build(),
-        title: this.test?.fullTitle(),
-      },
-      async ({ driver }) => {
-        await login(driver);
-
-        await new HeaderNavbar(driver).openContactsPage();
-        const contactsPage = new ContactsPage(driver);
-        await contactsPage.checkPageIsLoaded();
-        await contactsPage.addContactNewChain(
-          'Test User',
-          '0x56A355d3427bC2B1E22c78197AF091230919Cc2A',
-          'Sepolia',
-        );
-        await contactsPage.checkContactDisplayed({
-          contactName: 'Test User',
-          address: shortenAddress('0x56A355d3427bC2B1E22c78197AF091230919Cc2A'),
-        });
-      },
-    );
-  });
-
-  it('Edit entry in address book', async function () {
-    await withFixtures(
-      {
-        fixtures: new FixtureBuilderV2()
-          .withAddressBookController({
-            addressBook: {
-              '0x539': {
-                '0x2f318C334780961FB129D2a6c30D0763d9a5C970': {
-                  address: '0x2f318C334780961FB129D2a6c30D0763d9a5C970',
-                  chainId: '0x539',
-                  isEns: false,
-                  memo: '',
-                  name: 'Test Name 1',
-                },
-              },
-            },
-          })
-          .build(),
-        title: this.test?.fullTitle(),
-      },
-      async ({ driver }) => {
-        await login(driver);
-
-        await new HeaderNavbar(driver).openContactsPage();
-        const contactsPage = new ContactsPage(driver);
-        await contactsPage.checkPageIsLoaded();
-        await contactsPage.editContact({
-          existingContactName: 'Test Name 1',
-          newContactName: 'Test Name Edit',
-          newContactAddress: '0x74cE91B75935D6Bedc27eE002DeFa566c5946f74',
-        });
-        await contactsPage.checkContactDisplayed({
-          contactName: 'Test Name Edit',
-          address: shortenAddress('0x74cE91B75935D6Bedc27eE002DeFa566c5946f74'),
-        });
-      },
-    );
-  });
-  it('Deletes existing entry from address book', async function () {
-    await withFixtures(
-      {
-        fixtures: new FixtureBuilderV2()
-          .withAddressBookController({
-            addressBook: {
-              '0x539': {
-                '0x2f318C334780961FB129D2a6c30D0763d9a5C970': {
-                  address: '0x2f318C334780961FB129D2a6c30D0763d9a5C970',
-                  chainId: '0x539',
-                  isEns: false,
-                  memo: '',
-                  name: 'Test Name 1',
-                },
-              },
-            },
-          })
-          .build(),
-        title: this.test?.fullTitle(),
-      },
-      async ({ driver }) => {
-        await login(driver);
-
-        await new HeaderNavbar(driver).openContactsPage();
-        const contactsPage = new ContactsPage(driver);
-        await contactsPage.checkPageIsLoaded();
-        await contactsPage.deleteContact('Test Name 1');
-
-        // it checks if account is deleted
-        await contactsPage.checkContactDisplayed({
-          contactName: 'Test Name 1',
-          address: shortenAddress('0x2f318C334780961FB129D2a6c30D0763d9a5C970'),
-          shouldDisplay: false,
-        });
-      },
-    );
-  });
-
-  it('User can add same address contacts on different chains', async function () {
-    await withFixtures(
-      {
-        fixtures: new FixtureBuilderV2().build(),
-        title: this.test?.fullTitle(),
-      },
-      async ({ driver }) => {
-        await login(driver);
-
-        await new HeaderNavbar(driver).openContactsPage();
-        const contactsPage = new ContactsPage(driver);
-        await contactsPage.checkPageIsLoaded();
-        await contactsPage.addContact(
-          'Test User 1',
-          '0x56A355d3427bC2B1E22c78197AF091230919Cc2A',
-        );
-        await contactsPage.checkContactDisplayed({
-          contactName: 'Test User 1',
-          address: shortenAddress('0x56A355d3427bC2B1E22c78197AF091230919Cc2A'),
-        });
-
-        await contactsPage.addContactNewChain(
-          'Test User 2',
-          '0x56A355d3427bC2B1E22c78197AF091230919Cc2A',
-          'Sepolia',
-        );
-        await contactsPage.checkContactDisplayed({
-          contactName: 'Test User 2',
-          address: shortenAddress('0x56A355d3427bC2B1E22c78197AF091230919Cc2A'),
-        });
-      },
-    );
-  });
-});
