+++ test/e2e/tests/settings/ipfs-toggle.pw.spec.ts
@@ -1,0 +1,77 @@
+import { test as pwTest } from '@playwright/test';
+import { MockttpServer } from 'mockttp';
+import { E2E_DRIVER } from '../../constants';
+import { withFixtures } from '../../helpers';
+import FixtureBuilderV2 from '../../fixtures/fixture-builder-v2';
+import { SMART_CONTRACTS } from '../../seeder/smart-contracts';
+import HeaderNavbar from '../../page-objects/pages/header-navbar';
+import Homepage from '../../page-objects/pages/home/homepage';
+import NFTDetailsPage from '../../page-objects/pages/nft-details-page';
+import NftsTab from '../../page-objects/pages/home/nfts-tab';
+import PrivacySettings from '../../page-objects/pages/settings/privacy-settings';
+import SettingsPage from '../../page-objects/pages/settings/settings-page';
+import { login } from '../../page-objects/flows/login.flow';
+import { closeSettings } from '../../page-objects/flows/settings.flow';
+
+async function mockIPFSRequest(mockServer: MockttpServer) {
+  return [
+    await mockServer
+      .forGet(
+        'https://bafkreifvhjdf6ve4jfv6qytqtux5nd4nwnelioeiqx5x2ez5yrgrzk7ypi.ipfs.dweb.link/',
+      )
+      .thenCallback(() => ({ statusCode: 200 })),
+  ];
+}
+
+pwTest.describe('Settings', () => {
+  const smartContract = SMART_CONTRACTS.ERC1155;
+  pwTest(
+    'Shows nft default image when IPFS toggle is off and restore image once we toggle the ipfs modal',
+    async (
+      // eslint-disable-next-line no-empty-pattern
+      {},
+      testInfo,
+    ) => {
+      await withFixtures(
+        {
+          dappOptions: { numberOfTestDapps: 1 },
+          fixtures: new FixtureBuilderV2().withNftControllerERC1155().build(),
+          smartContract,
+          driverType: E2E_DRIVER.PLAYWRIGHT,
+          title: testInfo.titlePath.join(' '),
+          testSpecificMock: mockIPFSRequest,
+        },
+        async ({ driver }) => {
+          await login(driver, { validateBalance: false });
+
+          await new HeaderNavbar(driver).openSettingsPage();
+          const settingsPage = new SettingsPage(driver);
+          await settingsPage.checkPageIsLoaded();
+          await settingsPage.goToPrivacySettings();
+          const privacySettings = new PrivacySettings(driver);
+          await privacySettings.checkPageIsLoaded();
+          await privacySettings.toggleIpfsGateway();
+          await closeSettings(driver);
+          const homePage = new Homepage(driver);
+          await homePage.checkPageIsLoaded();
+
+          await homePage.goToNftTab();
+          const nftsTab = new NftsTab(driver);
+          await nftsTab.checkPageIsLoaded();
+          await nftsTab.clickNFTIconOnActivityList();
+          const nftDetailsPage = new NFTDetailsPage(driver);
+          await nftDetailsPage.checkPageIsLoaded();
+
+          // check for default image
+          await nftDetailsPage.checkNftDefaultImageIsDisplayed();
+
+          // toggle on ipfs to show image
+          await nftDetailsPage.showNftImage();
+
+          // should render image now
+          await nftDetailsPage.checkNftRenderedImageIsDisplayed();
+        },
+      );
+    },
+  );
+});
